+++ restaurant/model/order_item.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, Float
 from restaurant.model.base import Base
 from sqlalchemy.orm import relationship
 from restaurant.model.mixin import DateTimeMixin
@@ -11,8 +11,8 @@ class OrderItem(Base, DateTimeMixin):
     __tablename__ = 'order_item'
     id = Column(Integer, primary_key=True)
     quantity = Column(Integer, default=1, nullable=False)
-    unit_amount = Column(Integer, nullable=False)
-    total_amount = Column(Integer, default=0, nullable=False)
+    unit_amount = Column(Float, nullable=False)
+    total_amount = Column(Float, default=0, nullable=False)
     order_id = Column(ForeignKey('order.id'))
     item_id = Column(ForeignKey('item.id'))
